connect booking to payment

+++ src/main/java/com/uts/Online/Booking/App/DAO/PaymentDAO.java
@@ -5,10 +5,10 @@ import com.uts.Online.Booking.App.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PaymentDAO extends JpaRepository<Payment, Integer> {
-    List<Payment> findByPaymentId(Integer paymentId);
+public interface PaymentDAO extends JpaRepository<Payment, Long> {
+    List<Payment> findByPaymentId(Long paymentId);
 
-    List<Payment> findByBookingId(Integer bookingId);
+    List<Payment> findByBookingId(Long bookingId);
 
-    List<Payment> findByUserId(Integer userId);
+    List<Payment> findByUserId(Long userId);
 }
