+++ src/main/java/com/uts/Online/Booking/App/model/Player.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class Player extends User {
 
     @Column(name="payment_id")
-    private int paymentId;
+    private Long paymentId;
 
     @Column(name="address")
     private String address;
@@ -24,11 +24,11 @@ public class Player extends User {
         this.address = address;
     }
 
-    public int getPaymentId() {
+    public Long getPaymentId() {
         return paymentId;
     }
 
-    public void setPaymentId(int paymentId) {
+    public void setPaymentId(Long paymentId) {
         this.paymentId = paymentId;
     }
